add https://stackoverflow.com/a/26038979/4260616

+++ npm/scripts/build.js
@@ -82,7 +82,13 @@ function buildAssets(root, assets, includeVendor)
 
 			// On the first entry write the file
 			if (index == 0) {
-				fs.copyFileSync(file, root + '/' + asset.dest);
+				if (fs.statSync(file).isFile()) {
+					fs.copyFileSync(file, root + '/' + asset.dest);
+					console.log('File ' + file + ' copied to ' + root + '/' + asset.dest);
+				} else {
+					copyFolderRecursiveSync(file, root + '/' + asset.dest);
+					console.log('Folder  ' + file + ' copied to ' + root + '/' + asset.dest);
+				}
 				return;
 			}
 
@@ -142,6 +148,43 @@ function buildAssets(root, assets, includeVendor)
 			fs.unlinkSync(root + '/' + asset.dest);
 		}
 	});
+}
+
+function copyFileSync( source, target ) {
+
+    var targetFile = target;
+
+    // If target is a directory, a new file with the same name will be created
+    if ( fs.existsSync( target ) ) {
+        if ( fs.lstatSync( target ).isDirectory() ) {
+            targetFile = path.join( target, path.basename( source ) );
+        }
+    }
+
+    fs.writeFileSync(targetFile, fs.readFileSync(source));
+}
+
+function copyFolderRecursiveSync( source, target ) {
+    var files = [];
+
+    // Check if folder needs to be created or integrated
+    var targetFolder = path.join( target, path.basename( source ) );
+    if ( !fs.existsSync( targetFolder ) ) {
+        fs.mkdirSync( targetFolder );
+    }
+
+    // Copy
+    if ( fs.lstatSync( source ).isDirectory() ) {
+        files = fs.readdirSync( source );
+        files.forEach( function ( file ) {
+            var curSource = path.join( source, file );
+            if ( fs.lstatSync( curSource ).isDirectory() ) {
+                copyFolderRecursiveSync( curSource, targetFolder );
+            } else {
+                copyFileSync( curSource, targetFolder );
+            }
+        } );
+    }
 }
 
 module.exports = {
